+++ packages/growerp/bin/functions/functions.dart
@@ -1,6 +1,8 @@
 export 'create_chat_env.dart';
 export 'create_flutter_env.dart';
 export 'create_moqui_env.dart';
-export 'find_file_path.dart';
+export 'get_file_location.dart';
+export 'get_growerp_package_info.dart';
 export 'get_package_list.dart';
+export 'switch_package.dart';
 export 'update_app_settings.dart';