+++ packages/growerp/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.0-dev.5
+
+- added the 'growerp switchPackage' command
+
+## 0.1.0-dev.4
+
+- added message when pub.dev limit is reached.
+
+## 0.1.0-dev.3
+
+- fix model files.
+
 ## 0.1.0-dev.2
 
 - Created install command.